fix: windows support for loading files
On windows, there are quite a few specifics when it comes to resolving
a local file. This complexity stemmed from the desire to support
common, albeit invalid URI constructs for windows.

Overall, the original sin of this function has been to ignore the host
part of an URI, but now we can't really break this behavior.

* fixes #72
* refactored the local path pre-processing, so the windows-specifics
  appear more clearly
* fixed a few inconsistencies on windows file URIs and guarded against a panic

* test: refactored tests for the local file strategy, now table-driven and
  more systematic
* doc: added detailed documentation to LoadStrategy()
* ci: re-enacted tests on windows
* ci: temporarily muted past linting issues: relinting is deferred to another PR

Signed-off-by: Frederic BIDON <fredbi@yahoo.com>