fix(mangling): fixed name mangling when special chars are substituted
This PR fixes name mangling methods (ToJSONName, ToCommandName, etc)
for cases when a special character is expanded, like so:
"get$ref" should be jasonified as "getDollarRef".

The issue was that the extra space added was not trimmed before
name reconstruction, so in the example above, we get "getDollar Ref"
instead.

* contributes go-swagger/go-swagger#2821

Signed-off-by: Frederic BIDON <fredbi@yahoo.com>