feat(loading): HTTP or file loader configurable with LoadingOptions (#104)
* feat(loading): HTTP or file loader configurable with LoadingOptions

This PR deprecates package-level configurations for the HTTP or file
loader utility, by introducing a set of options.

It is a non-breaking change, as package-level variables are still
consumed as the default options.

Besides existing configurations (timeout, basic auth, custom headers),
we add the possibility to inject a custom http client.

For the local file loader, we add the possibility to inject a FS,
for example to be used with a dir-rooted FS or an embedded FS.

* contributes https://github.com/go-openapi/loads/issues/72
* refactored and augmented loading tests, with t.Run for improved
  test readability

Signed-off-by: Frederic BIDON <fredbi@yahoo.com>

Signed-off-by: Frederic BIDON <fredbi@yahoo.com>

* Update loading.go

Co-authored-by: ccoVeille <3875889+ccoVeille@users.noreply.github.com>

---------

Signed-off-by: Frederic BIDON <fredbi@yahoo.com>
Co-authored-by: ccoVeille <3875889+ccoVeille@users.noreply.github.com>