Mangling: added support for pluralized initialisms
Pluralized initialisms are of the simplest form, e.g. ID -> IDs.

Initialisms that end with an S (such as DNS) or that conflict with
another initialism that ends with an S (such as HTTP) are invariant
and do not support a pluralized form.

* fixes #46

Other changes that stem from newly supported initialisms that are not
fully upper-cased:

* replaced IPV4 and IPV6 by IPv4 and IPv6
* now ToGoName doesn't force names like "IPv4Etc" in "IPV4Etc", but
  leaves the casing of IPv4.
* if it happens (by adding non default initialisms) that 2 initialisms
  differ only in case, the mixed-case one will be favored over the
  capitalized.

Signed-off-by: Frederic BIDON <fredbi@yahoo.com>